MAYA-112769 move selectability code
Move the code and metadata declarations in the utils folder. Use correct ifdef in the header file.
