+++ lib/mayaUsd/utils/plugInfo.json
